+++ trabalho-pratico/src/query_6.c
@@ -36,7 +36,7 @@ char *query_6(char *city, char *dateA, char *dateB, UserData *userData, DriverDa
 	//  Green: Tarifa mínima = 4.00€ + 0.79€/km
 	//  Premium: Tarifa mínima = 5.20€ + 0.94€/km
 
-	float avgDistance = (float)distance / (float)total;
+	double avgDistance = (double)distance / (double)total;
 
 	char *resultTruncated = malloc(STR_BUFF_SIZE * sizeof(char));
 	snprintf(resultTruncated, STR_BUFF_SIZE, "%.3f\n", avgDistance);
